# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ANPE CENTRE NORD API",
        "description": "Bienvenue dans la documentation de l'API de l'ANPE CENTRE NORD.<br>Cette API fournit un accès sécurisé aux fonctionnalités du backend, <br>permettant la gestion efficace des ressources liées à l'emploi, aux utilisateurs et aux données associées.",
        "contact": {
            "name": "SYMBOL SARL",
            "email": "admin@admin.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur test de l'API"
        }
    ],
    "paths": {
...
        "/api/profiles": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Liste des profiles",
                "description": "Retourne la liste des profiles",
                "operationId": "43d12304c3e621ecce8cc9ccd4b3738e",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Création d'un profile",
                "description": "Crée un nouveau profile et retourne le profile créé",
                "operationId": "00abb5f24a20112b161d17e9a5c38a37",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "label"
                                ],
                                "properties": {
                                    "apropos": {
                                        "type": "string",
                                        "example": "Passioné par le digital et les nouvelles technologies, je recherche des missions challengantes"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john.doe@example.com"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "75000000"
                                    },
                                    "ville": {
                                        "type": "string",
                                        "example": "Kaya"
                                    },
                                    "adresse": {
                                        "type": "string",
                                        "example": "Secteur 05"
                                    },
                                    "annee_experience": {
                                        "type": "string",
                                        "example": "2 ans"
                                    },
                                    "niveau_etude": {
                                        "type": "string",
                                        "example": "BAC + 3"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "profile créé avec succès"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Les données fournies ne sont pas valides"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profiles/{slug}": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Récupère un profile par son slug",
                "description": "Retourne un profile",
                "operationId": "d35ac6741337ef9d18e65903752625c2",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "slug du profile à récupérer",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Modification d'un profile",
                "description": "Modifie un profile et retourne le profile modifié",
                "operationId": "13272c76ae703dcfaec26d3e2cdffe77",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "slug du profile à modifié",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "label"
                                ],
                                "properties": {
                                    "apropos": {
                                        "type": "string",
                                        "example": "Passioné par le digital et les nouvelles technologies, je recherche des missions challengantes"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john.doe@example.com"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "75000000"
                                    },
                                    "ville": {
                                        "type": "string",
                                        "example": "Kaya"
                                    },
                                    "adresse": {
                                        "type": "string",
                                        "example": "Secteur 05"
                                    },
                                    "annee_experience": {
                                        "type": "string",
                                        "example": "2 ans"
                                    },
                                    "niveau_etude": {
                                        "type": "string",
                                        "example": "BAC + 3"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "profile modifié avec succès"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Slug validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "profile non trouvé"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Les données fournies ne sont pas valides"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Profile"
                ],
                "summary": "Supprime un profile par son slug",
                "description": "Retourne le profile supprimé",
                "operationId": "51ee2a727a72327fb70831f4fc050808",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "slug du profile à supprimer",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "'profile supprimé avec succès"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Slug validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Profile non trouvé"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}